sql-performance-monitoring-tool/sql-performance-monitoring-tool/requirements.txt
sqlalchemy
pandas
matplotlib
seaborn
psycopg2-binary
numpy
pyyaml